+++ src/input/textarea.ts
@@ -4,11 +4,11 @@ import { IPropWidget } from "../interface/widget";
 
 import { textareaCls } from "../theme";
 import { setValue, titleFromLabel } from "../utils";
-import { ValidationBase } from "../validationBase";
 
+import { ValidationBase } from "../validationBase";
 import { LayoutTop } from "./layout/layoutTopLabel";
 
-export class TextareaInput extends ValidationBase {
+export class TextareaInput extends ValidationBase<IPropWidget> {
 
 	protected override readonly selector = "textarea";
 
@@ -20,16 +20,19 @@ export class TextareaInput extends ValidationBase {
 			minlength, maxlength, tabindex, spellcheck,
 			instant, uiClass = {}
 		} = attrs.field;
-		return m(LayoutTop, { field, value, xform, invalid: this.invalid },
-			m("textarea.w-100.bg-transparent.bn.outline-0.h-100.resize-none", {
-				id, name, title,
-				placeholder, required, readonly, disabled, autofocus, autocomplete,
-				minlength, maxlength, tabindex, spellcheck,
-				class: textareaCls(uiClass),
-				value: value(),
-				// Update value on change or input ("instant" option)
-				[instant ? "oninput" : "onchange"]: setValue(value)
-			})
-		);
+		return m(LayoutTop, {
+			field,
+			value,
+			xform,
+			invalid: this.invalid
+		}, m("textarea.w-100.bg-transparent.bn.outline-0.h-100.resize-none", {
+			id, name, title,
+			placeholder, required, readonly, disabled, autofocus, autocomplete,
+			minlength, maxlength, tabindex, spellcheck,
+			class: textareaCls(uiClass),
+			value: value(),
+			// Update value on change or input ("instant" option)
+			[instant ? "oninput" : "onchange"]: setValue(value)
+		}));
 	}
 }